+++ server/data/reddit.js
@@ -1,30 +1,43 @@
 // Reddit subs data fetcher
 
 import request from 'request';
+import moment from 'moment';
 import redditModel from '../models/Reddit';
-//const snoowrap = require('snoowrap');
+
+const location = 'Worldwide';
 
 const redditSubs = {
+    location: 'Worldwide',
     getSubs: () => {
-        const location = 'Worldwide';
-        request({url: 'https://www.reddit.com/api/trending_subreddits.json', json: true}, (error, response, body) => {
-            if (error || response.statusCode !== 200) {
-                console.log(`An unexpected error occured while requesting Reddit Subs.`);
+        redditModel.findOne({}, {}, { sort: { 'created_at' : -1 } }, (err, item) => {
+            let getData = false;
+            if (err || !item) {
+                getData = true;
             }
             else {
-                if (body) {
-                    console.log('[trends server] Updating all database entries for Reddit Subs (Worldwide)');
-                    /*const r = new snoowrap({
-                        userAgent: 'put your user-agent string here',
-                        clientId: 'put your client id here',
-                        clientSecret: 'put your client secret here',
-                        refreshToken: 'put your refresh token here'
-                    });*/
-                    redditModel.deleteMany((error) => {
-                        if (error) return console.error(error);
-                        redditSubs.insertSubs(body);
-                    });
+                if (!moment(item.createdAt).isAfter(moment().subtract(1, 'hours'))) {
+                    getData = true;
                 }
+            }
+            if (getData) {
+                console.log(`[trends server] Database entries for Reddit Subs (${redditSubs.location}) are already up to date.`);
+                request({url: 'https://www.reddit.com/api/trending_subreddits.json', json: true}, (error, response, body) => {
+                    if (error || response.statusCode !== 200) {
+                        console.log(`An unexpected error occured while requesting Reddit Subs.`);
+                    }
+                    else {
+                        if (body) {
+                            console.log(`[trends server] Updating all database entries for Reddit Subs (${redditSubs.location})`);
+                            redditModel.deleteMany((error) => {
+                                if (error) return console.error(error);
+                                redditSubs.insertSubs(body);
+                            });
+                        }
+                    }
+                }); 
+            }
+            else {
+                console.log(`[trends server] Database entries for Reddit Subs (${redditSubs.location}) are already up to date.`);
             }
         });
     },
@@ -47,7 +60,7 @@ const redditSubs = {
                         }
                         const dbData = new redditModel({
                             name: sub,
-                            location: 'Worldwide',
+                            location: redditSubs.location,
                             url: `https://reddit.com/r/${sub}`,
                             subscribers: subscribers
                         });
